pyproject.toml
[project]
name = "ruvon-echoforge"
version = "0.1.1"
description = "EchoForge Syndicate — sovereign distributed quant intelligence mesh for the Ruvon ecosystem"
authors = [{name = "Ruvon Team", email = "team@ruvon.dev"}]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.10"
keywords = ["ruvon", "echoforge", "quant", "trading", "wasm", "webrtc", "edge"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "ruvon-sdk>=0.1.3",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=12.0",
    "pydantic>=2.0",
    "aiosqlite>=0.19.0",
    "nats-py>=2.6.0",
    "protobuf>=4.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "httpx>=0.27"]

[project.scripts]
echoforge = "ruvon_echoforge.bridge.main:cli"

[project.urls]
Homepage = "https://github.com/KamikaziD/ruvon-echoforge"
Repository = "https://github.com/KamikaziD/ruvon-echoforge"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ruvon_echoforge"]

[tool.hatch.build.targets.sdist]
include = [
    "src/",
    "tests/",
    "workflows/",
    "pyproject.toml",
    "README.md",
    "LICENSE",
]
